logmanager-project/requirements.txt
Django==4.2.7
psycopg2-binary==2.9.9
djangorestframework==3.14.0
pandas==2.1.1
numpy==1.26.0
celery==5.3.4
redis==5.0.1
django-celery-beat==2.5.0
django-celery-results==2.5.1
python-dotenv==1.0.0
django-filter==23.3
djangorestframework-simplejwt==5.3.0
channels==4.0.0
daphne==4.0.0
django-cors-headers==4.3.0
Pillow==10.1.0
matplotlib==3.8.0
plotly==5.18.0
nltk==3.8.1
scikit-learn==1.3.2 